code/requirements.txt
google-adk
requests
websocket-client
